+++ crates/wdl-lint/src/lib.rs
@@ -101,6 +101,8 @@ pub fn rules() -> Vec<Box<dyn Rule>> {
         Box::<rules::BlankLinesBetweenElementsRule>::default(),
         Box::<rules::KeyValuePairsRule>::default(),
         Box::<rules::ExpressionSpacingRule>::default(),
+        Box::<rules::DisallowedInputNameRule>::default(),
+        Box::<rules::DisallowedOutputNameRule>::default(),
     ];
 
     // Ensure all the rule ids are unique and pascal case